Move code into independent modules
