+++ src/frink_embeddings_web/core/errors.py
