+++ src/frink_embeddings_web/evaluation/templates/eval_report.md
