+++ src/frink_embeddings_web/web/__init__.py
@@ -1,0 +1,6 @@
+from .app import create_app
+
+
+def main() -> None:
+    app = create_app()
+    app.run()
