+++ src/frink_embeddings_web/web/app.py
@@ -1,7 +1,7 @@
 from loguru import logger
 
-from frink_embeddings_web.context import AppContext, WrappedFlask
-from frink_embeddings_web.settings import load_settings
+from ..config import AppContext, load_settings
+from ._flask import WrappedFlask
 
 
 def create_app() -> WrappedFlask:
@@ -15,7 +15,7 @@ def create_app() -> WrappedFlask:
     app.ctx = ctx
 
     # Register API & Web routes
-    from frink_embeddings_web.routes import api, web
+    from .routes import api, web
 
     app.register_blueprint(api)
     app.register_blueprint(web)
