+++ src/frink_embeddings_web/web/static/index.js
