+++ src/frink_embeddings_web/web/static/style.css
