+++ src/frink_embeddings_web/web/templates/index.html
