+++ src/frink_embeddings_web/web/templates/partials/results_table.html
